// src/lib/reward-api/models/data_ApproveFinanceDocRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type data_ApproveFinanceDocRequest = {
    remark?: string;
    status: data_ApproveFinanceDocRequest.status;
};
export namespace data_ApproveFinanceDocRequest {
    export enum status {
        APPROVED = 'APPROVED',
        REJECTED = 'REJECTED',
    }
}