// src/lib/reward-api/models/data_SubmitIssueRequestRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type data_SubmitIssueRequestRequest = {
    remark?: string;
    status: data_SubmitIssueRequestRequest.status;
};
export namespace data_SubmitIssueRequestRequest {
    export enum status {
        TO_APPROVE = 'TO_APPROVE',
    }
}